// generated/src/app/api-models/pmb-subsc-hist.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class PmbSubscHist {
  customerType: string;
  customerId: string;
  invoiceNo: number;
  subscriberId: string;
  personNumber: string;
  seqGpbilId: number;
  groupId: string;
  employeeNo: string;
  subscLastName: string;
  subscFirstName: string;
  subscDateOfBirth: Date;
  subscGender: string;
  subscAge: number;
  subscDept: string;
  subscLocation: string;
  subscSalary: number;
  familySize: number;
  noOfChildDep: number;
  noOfAdultDep: number;
  premiumStep: string;
  medicareStatus: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  rateType: string;
  spouseFlag: string;
  otherStatusFlag: string;
  premOverrideStep: number;
  premOverrideAmt: number;
  premOverrideCode: string;
  cobraFlag: string;
  monthlyRate: number;
  seqParentId: number;
}